Implements SceneObjects, Lights, Materials
Still currently only a hard-coded sphere as the scene, but now that
sphere is wrapped in a "SceneObject" with a "Material" subclass to
specify how to shade the sphere. Available materials are NormalMaterial
(which just uses the normal with a transformation as the color),
PositionMaterial (which just uses the position with a multiplier as the
color), and BlinnPhongMaterial.

Also implements a simple point light for lighting.

Also many changes to the math library to fill out functionality and
operators.