Refactors to remove float/double template concept
This commit refactors all of the code to remove the non-useful
'std::floating_point T' template on most of the code. While it was nice
that in theory all of the code could work with either float or double,
it prevented moving code into cpp files and made some things a pain. It
was also unnecessary: float is enough precision, and would be required
if I ever want to port the code to run on the GPU, since 32 bit floats
are the basic unit in GPUs.